Renaming packages + renaming TableConstants
