+++ app/src/main/java/ayds/winchester/songinfo/moredetails/data/card/repository/local/card/sqldb/CardTableConsants.kt
@@ -1,4 +1,4 @@
-package ayds.winchester.songinfo.moredetails.data.wikipedia.repository.local.wikipedia.sqldb
+package ayds.winchester.songinfo.moredetails.data.card.repository.local.card.sqldb
 
 const val COLUMN_ID = "id"
 const val COLUMN_ARTIST = "artist"
@@ -9,7 +9,6 @@ const val COLUMN_SOURCE = "source"
 const val TABLE_NAME = "artists"
 const val COLUMNS_FOR_WHERE = "$COLUMN_ARTIST = ?"
 const val SORT_ORDER_CURSOR = "$COLUMN_ARTIST DESC"
-const val VALUE_SOURCE = 1
 const val SQL_CREATE_TABLE =
     "create table $TABLE_NAME (" +
             "$COLUMN_ID INTEGER PRIMARY KEY AUTOINCREMENT," +
